Add files via upload
The Loan Payment dataset contains above mentioned attributes. We would like to perform classification on Loan_status variable that contains three categories namely 	PAIDOFF,COLLETION AND COLLECTION_PAIDOFF. The PAIDOFF category is all about loan paid off by the customers. The COLLECTION is all about the loan is yet to be paid off so bank is collecting money from those customers who seem to be unwilling to repay the loan amount; these customers are apprehended to be defaulters.
The next and last category is COLLECTION_PAIDOFF that states that the loan amount was repaid by the customers after conducting Collection drive. So we’ve all other variables available to classify customers into any of the given categories.
This analysis will generate valuable insights for banks to consider which customers they will provide loan to.